feat(terminal-types): github-maintenance plugin — per-repo tab
Closes bd remote-dev-1ebu.4.
