+++ src/lib/terminal-plugins/init-client.ts
@@ -21,6 +21,7 @@ import { ProfilesClientPlugin } from "./plugins/profiles-plugin-client";
 import { PortManagerClientPlugin } from "./plugins/port-manager-plugin-client";
 import { TrashClientPlugin } from "./plugins/trash-plugin-client";
 import { ProjectPrefsClientPlugin } from "./plugins/project-prefs-plugin-client";
+import { GitHubMaintenanceClientPlugin } from "./plugins/github-maintenance-plugin-client";
 
 // Client-only: use console directly to avoid pulling the server-side logger
 // (which depends on better-sqlite3) into the browser bundle.
@@ -58,6 +59,7 @@ export function initializeClientPlugins(): void {
   TerminalTypeClientRegistry.register(PortManagerClientPlugin, { builtIn: true });
   TerminalTypeClientRegistry.register(TrashClientPlugin, { builtIn: true });
   TerminalTypeClientRegistry.register(ProjectPrefsClientPlugin, { builtIn: true });
+  TerminalTypeClientRegistry.register(GitHubMaintenanceClientPlugin, { builtIn: true });
 
   TerminalTypeClientRegistry.setDefaultType("shell");
